requirements.txt
openai>=1.30.0
pydantic==2.8.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-dotenv==1.0.1
fire==0.3.1
httpx>=0.27.0
google-api-python-client
